requirements.txt
#Template
black==22.3.0
mypy==0.942
pylint==2.13.4
pytest-cov==3.0.0
pytest-xdist==2.5.0
pytest==7.0.1
pyaml-env==1.2.1
#Project
PySide2==5.15.2.1
Pillow==9.3.0
requests==2.7.0